+++ packages/nimble-components/src/select/styles.ts
@@ -37,6 +37,13 @@ export const styles = css`
         overflow-y: auto;
         position: absolute;
         width: 100%;
+        ${
+            /*
+             * The --max-height custom property is defined by fast-foundation.
+             * See: https://github.com/microsoft/fast/blob/af1f8736ae9ff54a7449324bad952865981d1ece/packages/web-components/fast-foundation/src/select/select.ts#L199
+             */ ''
+        }
+        max-height: calc(var(--max-height) - ${controlHeight});
         z-index: 1;
         padding: 4px;
         box-shadow: 0px 3px 3px ${popupBoxShadowColor};